requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dateutils==0.6.12
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.102.0
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
nsetools==1.0.12
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.20.0


fastapi==0.102.0
python-dotenv==1.0.0
youtube-transcript-api==0.6.3
uvicorn==0.20.0
groq==0.18.0